Fix R2 score calculation issue in regression model by simplifying tensor shape handling

+++ src/models/networks/heads.py
@@ -13,5 +13,4 @@ class ClsRegHead(nn.Module):
         x = self.global_pool(x)
         x = torch.flatten(x, 1)
         x = self.fc(x)
-        # Ensure output is float for regression tasks
-        return x.float()
+        return x
